extract nhwc
